Reimbursement-Application/src/test/java/com/shane/controller/UserControllerTest.java
package com.shane.controller;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import javax.servlet.http.HttpSession;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.shane.model.Response;
import com.shane.model.Users;
import com.shane.service.UserService;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;

import java.io.IOException;
import java.io.PrintWriter;
import java.io.StringWriter;
import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;


class UserControllerTest extends Mockito {

    ////I could not get any of my Controller test to work due to Request and Response Issues


    @Mock
    private UserController userController;
    @Mock
    private UserService userService;



    @BeforeEach
    void setUp() throws Exception {
        MockitoAnnotations.initMocks(this);

    }

    @Test
    void getAllUsers() throws IOException {
/*        HttpServletRequest request = mock(HttpServletRequest.class);
        HttpServletResponse response = mock(HttpServletResponse.class);


        when(request.getParameter("username")).thenReturn("AlbaO");
        response.setContentType("application/json");

        StringWriter stringWriter = new StringWriter();
        PrintWriter writer = new PrintWriter(stringWriter);



        when(userService.getAllEmployees()).thenReturn(expectedResult);
        when(response.getWriter()).thenReturn(writer);
        userController.getAllUsers(request, response);
        verify(userController, atLeast(1));
        assertEquals(request, expectedResult);*/
    }
    @Test
    public void login() throws ServletException, IOException {
/*        HttpServletRequest request = mock(HttpServletRequest.class);
        HttpServletResponse response = mock(HttpServletResponse.class);
        HttpSession session = mock(HttpSession.class);


        session.setAttribute("username", "AlbaO");
        session.setAttribute("password", "OpenUp");
        when(request.getSession()).thenReturn(session);
        // Get parameters
        when(request.getParameter("username")).thenReturn("AlbaO");
        when(request.getParameter("password")).thenReturn("OpenUp");

        StringWriter stringWriter = new StringWriter();
        PrintWriter writer = new PrintWriter(stringWriter);


        when(response.getWriter()).thenReturn(writer);

        userController.login(request, response);

        verify(request, atLeast(1)).getContentType();
        verify(response, atLeast(1)).getContentType();*/

        //verify(request, atLeast(1)).getParameter("username");

        //assertEquals(request.toString(),response.toString());
/*        Users users = new Users(
                "AlbaO",
                "OpenUp",
                "Alba",
                "Orozco",
                "albao@email.com",
                2);
        HttpServletRequest req = mock(HttpServletRequest.class);
        HttpServletResponse resp = mock(HttpServletResponse.class);
        HttpSession session = mock(HttpSession.class);

        session.setAttribute("username", "AlbaO");
        session.setAttribute("password", "OpenUp");

        // Get parameters
        when(req.getParameter("username")).thenReturn("AlbaO");
        when(req.getParameter("password")).thenReturn("OpenUp");

        when(session.getAttribute("username")).thenReturn("AlbaO");
        when(session.getAttribute("password")).thenReturn("OpenUp");

        StringWriter stringWriter = new StringWriter();
        PrintWriter writer = new PrintWriter(stringWriter);

        when(resp.getWriter()).thenReturn(writer);

        Mockito.when(userController.login(req, resp)).getClass(users);

        verify(req, atLeast(1)).getParameter("username");

        writer.flush();
        assertTrue(resp.toString().contains("username"));*/

    }

    @Test
    void addUser() throws Exception, IOException {
/*        Users users = new Users(
                "AlbaO",
                "OpenUp",
                "Alba",
                "Orozco",
                "albao@email.com",
                2);
        when(request.getParameter("username")).thenReturn("AlbaO");
        when(request.getParameter("password")).thenReturn("OpenUp");

        StringWriter sw = new StringWriter();
        PrintWriter pw = new PrintWriter(sw);
       // Mockito.when(response.getWriter().append("username: " + users.getUsername() + "password: " + users.getPassword()));


        when(response.getWriter()).thenReturn(pw);


        userController.addUser(request, response);
        String result = sw.getBuffer().toString().trim();
        String acutal = "username: AlbaO password: OpenUp";
        assertEquals(result, acutal);*/


/*      StringWriter stringWriter = new StringWriter();
       PrintWriter writer = new PrintWriter(stringWriter);

       Mockito.when(request.getParameter("username")).thenReturn("AlbaO");
       Mockito.when(response.getWriter()).thenReturn(writer);

        //Assign
        Users users = new Users(
                "AlbaO",
                "OpenUp",
                "Alba",
                "Orozco",
                "albao@email.com",
                2);
        Users tempUser = users;

       doNothing().when(userController).addUser(request, response);
       userController.addUser(request, response);

        verify(userController, times(1)).addUser(request, response);*/



       /* Users expectResult = when(request.getParameter("username")).thenReturn("AlbaO");
        //String parameter = request.getParameter(("username").thenReturn("AlbaO"));
        when(request.getParameter("username")).thenReturn("AlbaO");
        when(request.getParameter("password")).thenReturn("OpenUp");
        Users temporaryUser = users;

        when(response.getWriter()).thenReturn(writer);
        Mockito.when(userController.addUser(request, response)).thenReturn("AlbaO","OpenUp");*/
    /*    StringWriter stringWriter = new StringWriter();
        PrintWriter writer = new PrintWriter(stringWriter);
        when(response.getWriter()).thenReturn(writer);
        when(userController.addUser(request, response)).thenReturn(T);

        Mockito.when(userController.addUser(request, response)).thenReturn(responses);
        assertEquals("text/html", response.getContentType());*/



/*        when(request.getParameter(users.toString())).thenReturn(users.toString());

        StringWriter stringWriter = new StringWriter();
        PrintWriter writer = new PrintWriter(stringWriter);
        when(response.getWriter()).thenReturn(writer);

        userController.addUser(request, response);
        verify(request, atLeast(1)).getParameter(users.toString());
        writer.flush();
        assertTrue(stringWriter.toString().contains(users.toString()));
        */



/*        String result = userController.addUser(HttpServletRequest req, HttpServletResponse resp);
        //given(userService.addUser()).willAnswer((invocation)-> invocation.getArgument(0))
        //Act
        Mockito.when(userController.addUser()).thenReturn(expectResult);
        userController.addUser(users);

        //Assert
        verify(userDao, times(1)).createUser(users);*/
    }

    @Test
    void checkSession() {
/*        HttpServletRequest request = mock(HttpServletRequest.class);
        HttpServletResponse response = mock(HttpServletResponse.class);
        HttpSession session = mock(HttpSession.class);

        public String check()
        session.setAttribute("username", "AlbaO");
        session.setAttribute("password", "OpenUp");
        Users users = new Users(
                "AlbaO",
                "OpenUp",
                "Alba",
                "Orozco",
                "albao@email.com",
                2);
        String users1 = users.toString();
        Users sessionUser = (Users) session.getAttribute(users1);


        when(request.getSession()).thenReturn(session);
        // Get parameters
        when(request.getParameter("username")).thenReturn("AlbaO");
        when(request.getParameter("password")).thenReturn("OpenUp");

        response.setContentType("application/json");*/
        //if(create && (session))
    }

/*    let createNew = await fetch(`${domain}/api/reimbursement`,{
        method: "POST",
                body: JSON.stringify({
                reimburAmount: conVert,
                reimburDescription: newReimburDesri,
                reimburAuthorFk: userId,
                reimburTypeFk: value*/
    @Test
    void logOut() {
    }
}